requirements.txt
python-dotenv
docker